refactor: consolidate undefined option tests
